backend/requirements.txt
fastapi
uvicorn
pandas
requests
python-multipart
openpyxl
sqlalchemy
pymysql
python-jose[cryptography]
bcrypt==4.0.1
passlib==1.7.4
python-dotenv
